+++ src/gwt/src/org/rstudio/studio/client/server/remote/ClientEvent.java
@@ -206,6 +206,7 @@ class ClientEvent extends JavaScriptObject
    public static final String DeploymentRecordsUpdated = "deployment_records_updated";
    public static final String FormatDocumentCompleted = "format_document_completed";
    public static final String RunAutomation = "run_automation";
+   public static final String CopilotStatusChanged = "copilot_status_changed";
    
    protected ClientEvent()
    {