+++ src/gwt/src/org/rstudio/studio/client/workbench/views/packages/ui/PackageLinkColumn.java
@@ -42,9 +42,7 @@ import com.google.gwt.safehtml.shared.SafeUri;
 import com.google.gwt.user.cellview.client.Column;
 import com.google.gwt.view.client.ListDataProvider;
 
-import jsinterop.base.Any;
 import jsinterop.base.Js;
-import jsinterop.base.JsPropertyMap;
 
 
 // package name column which includes a hyperlink to package docs
@@ -64,7 +62,10 @@ public abstract class PackageLinkColumn extends Column<PackageInfo, PackageInfo>
                             final OperationWithInput<PackageInfo> onClicked,
                             final boolean alwaysUnderline)
    {
-      super(new AbstractCell<PackageInfo>("click")
+      public PackageLinkCell(final ListDataProvider<PackageInfo> dataProvider,
+                             final RepositoryPackageVulnerabilityListMap vulns,
+                             final OperationWithInput<PackageInfo> onClicked,
+                             final boolean alwaysUnderline)
       {
          // render anchor using custom styles. detect selection and
          // add selected style to invert text color
@@ -84,70 +85,89 @@ public abstract class PackageLinkColumn extends Column<PackageInfo, PackageInfo>
             }
          }
 
-         private void addVulnerabilityInfo(Context context, PackageInfo value, SafeHtmlBuilder sb)
+         sb.append(NAME_TEMPLATE.render(classNames, value.getName()));
+      }
+
+      private void addVulnerabilityInfo(Context context, PackageInfo value, SafeHtmlBuilder sb)
+      {
+         final Mutable<Boolean> didFindVulnerability = new Mutable<>(false);
+
+         String name = value.getName();
+         String version = value.getVersion();
+
+         vulns_.forEach((String key) ->
          {
-            if (vulns == null)
+            PackageVulnerabilityListMap pvlMap = vulns_.get(key);
+            if (pvlMap == null || !pvlMap.has(name))
                return;
-            
-            final Mutable<Boolean> didFindVulnerability = new Mutable<>(false);
 
-            String name = value.getName();
-            String version = value.getVersion();
-
-            vulns.forEach((String key) ->
+            PackageVulnerabilityList pvList = pvlMap.get(name);
+            for (int i = 0, n = pvList.getLength(); i < n; i++)
             {
-               PackageVulnerabilityListMap pvlMap = Js.uncheckedCast(vulns.get(key));
-               if (pvlMap == null || !pvlMap.has(name))
-                  return;
-               
-               PackageVulnerabilityList pvList = Js.uncheckedCast(pvlMap.get(name));
-               for (PackageVulnerability pvItem : pvList.asList())
+               PackageVulnerability pvItem = pvList.getAt(i);
+               if (pvItem.versions.has(version))
                {
-                  if (pvItem.versions.has(version))
-                  {
-                     SafeUri uri = RESOURCES.iconWarning().getSafeUri();
-                     String title = pvItem.id + ": " + pvItem.summary + "\n\n" + pvItem.details;
-                     sb.append(ICON_TEMPLATE.render(RESOURCES.styles().icon(), title, uri));
-                     didFindVulnerability.set(true);
-                     return;
-                  }
+                  SafeUri uri = RESOURCES.iconWarning().getSafeUri();
+                  String title = pvItem.id + ": " + pvItem.summary + "\n\n" + pvItem.details;
+                  sb.append(ICON_TEMPLATE.render(RESOURCES.styles().icon(), title, uri));
+                  didFindVulnerability.set(true);
+                  return;
                }
-            });
-
-            if (!didFindVulnerability.get())
-            {
-               sb.append(NONE_TEMPLATE.render(RESOURCES.styles().icon()));
             }
-         }
+         });
 
-         // click event which occurs on the actual package link div
-         // results in showing help for that package
-         @Override
-         public void onBrowserEvent(Context context, Element parent,
-                                    PackageInfo value, NativeEvent event,
-                                    ValueUpdater<PackageInfo> valueUpdater)
+         if (!didFindVulnerability.get())
          {
-            super.onBrowserEvent(context, parent, value, event, valueUpdater);
-            if ("click".equals(event.getType()))
+            sb.append(NONE_TEMPLATE.render(RESOURCES.styles().icon()));
+         }
+      }
+
+      // click event which occurs on the actual package link div
+      // results in showing help for that package
+      @Override
+      public void onBrowserEvent(Context context, Element parent, PackageInfo value,
+                                 NativeEvent event, ValueUpdater<PackageInfo> valueUpdater)
+      {
+         super.onBrowserEvent(context, parent, value, event, valueUpdater);
+         if ("click".equals(event.getType()))
+         {
+            // verify that the click was on the package link
+            JavaScriptObject target = event.getEventTarget().cast();
+            if (!Element.is(target))
+               return;
+
+            Element targetEl = Element.as(target);
+            if (!targetEl.hasClassName(RESOURCES.styles().link()))
+               return;
+
+            int idx = context.getIndex();
+            List<PackageInfo> data = dataProvider_.getList();
+            if (idx >= 0 && idx < dataProvider_.getList().size())
             {
-               // verify that the click was on the package link
-               JavaScriptObject target = event.getEventTarget().cast();
-               if (!Element.is(target))
-                  return;
-               
-               Element targetEl = Element.as(target);
-               if (!targetEl.hasClassName(RESOURCES.styles().link()))
-                  return;
-               
-               int idx = context.getIndex();
-               List<PackageInfo> data = dataProvider.getList();
-               if (idx >= 0 && idx < dataProvider.getList().size())
-               {
-                  onClicked.execute(data.get(idx));
-               }
+               onClicked_.execute(data.get(idx));
             }
          }
-      });
+      }
+
+      private final ListDataProvider<PackageInfo> dataProvider_;
+      private final RepositoryPackageVulnerabilityListMap vulns_;
+      private final OperationWithInput<PackageInfo> onClicked_;
+      private final boolean alwaysUnderline_;
+   }
+
+   public PackageLinkColumn(ListDataProvider<PackageInfo> dataProvider,
+                            RepositoryPackageVulnerabilityListMap vulns,
+                            OperationWithInput<PackageInfo> onClicked)
+   {
+      this(dataProvider, vulns, onClicked, false);
+   }
+
+   public PackageLinkColumn(final ListDataProvider<PackageInfo> dataProvider,
+                            final RepositoryPackageVulnerabilityListMap vulns,
+                            final OperationWithInput<PackageInfo> onClicked,
+                            final boolean alwaysUnderline)
+   {
+      super(new PackageLinkCell(dataProvider, vulns, onClicked, alwaysUnderline));
    }
 
    interface NameTemplate extends SafeHtmlTemplates
@@ -171,7 +191,9 @@ public abstract class PackageLinkColumn extends Column<PackageInfo, PackageInfo>
    interface Styles extends CssResource
    {
       String icon();
+
       String link();
+
       String linkUnderlined();
    }